requirements.txt
dj-database-url==0.2.2
django-extensions==1.4.0
django-model-utils==2.2
Django==1.7.1
djangorestframework==3.0.2
gevent==1.0.1
greenlet==0.4.4
gunicorn==19.1.1
ipdb==0.8
ipython==2.0.0
jsonfield==1.0.0
newrelic==2.20.0.17
psycogreen==1.0
psycopg2==2.5.1
